[Bug fix] The "RUN AUTOGRADER" button can cause database deadlock
The request handler for the "RUN AUTOGRADER" button is one function where two
transactions are used in the same execution frame. Because context manager
variables stay in the local scope until the function returns, the second cursor
is instantiated while the first one remains active. In certain circumstances,
this can cause a permanent deadlock condition where the database becomes
unavailable for all requests.

In order to address this issue, we must not allow database users to keep a copy
of APSW's cursor object in their local execution frame. The DbCursor object now
holds the only reference to the cursor object, so the destructor can be called
immediately after the transaction is finished.